+++ hocr_tools_lib/tools/hocr_eval.py
@@ -156,12 +156,12 @@ def main() -> None:
     parser = argparse.ArgumentParser()
     parser.add_argument(
         "truth", help="hOCR file with ground truth",
-        type=argparse.FileType("r")
+        type=str
     )
     parser.add_argument(
         "actual",
         help="hOCR file from the actual recognition",
-        type=argparse.FileType("r")
+        type=str
     )
     parser.add_argument("-d", "--debug", action="store_true")
     parser.add_argument("-v", "--verbose", action="store_true")
@@ -173,7 +173,7 @@ def main() -> None:
     #     "-o", "--significant_overlap", type=float, default=0.1,
     #     help="default: %(default)s"
     # )
-    parser.add_argument("-i", "--imgfile", type=argparse.FileType("r"))
+    parser.add_argument("-i", "--imgfile", type=str)
     args = parser.parse_args()
 
     image, segmentation_errors, segmentation_ocr_errors, ocr_errors = evaluate(
@@ -187,6 +187,3 @@ def main() -> None:
 
     if image:
         image.show("errors.png")
-
-    args.truth.close()
-    args.actual.close()