+++ tests/tools/test_hocr_pdf.py
@@ -17,22 +17,31 @@ class HocrPdfTestCase(TestCase):
     BASE_URL = "https://digi.bib.uni-mannheim.de/fileadmin/digi/445442158"
     WORK = "445442158_0126"
 
-    def _download_files(self, directory: Path, *, skip_hocr: bool = False) -> None:
+    hocr_data = b""
+    jpg_data = b""
+
+    @classmethod
+    def setUpClass(cls) -> None:
+        super().setUpClass()
+
+        with requests.Session() as session:
+            response = session.get(f"{cls.BASE_URL}/tess/{cls.WORK}.hocr")
+            assert response.status_code == 200, response.status_code
+            cls.hocr_data = response.content
+
+            response = session.get(f"{cls.BASE_URL}/max/{cls.WORK}.jpg")
+            assert response.status_code == 200, response.status_code
+            cls.jpg_data = response.content
+
+    def _generate_files(self, directory: Path, *, skip_hocr: bool = False) -> None:
         hocr_file = directory / f"{self.WORK}.hocr"
         jpg_file = directory / f"{self.WORK}.jpg"
 
-        session = requests.Session()
-        self.addCleanup(session.close)
-
         if not skip_hocr:
-            response = session.get(f"{self.BASE_URL}/tess/{self.WORK}.hocr")
-            self.assertEqual(200, response.status_code)
-            hocr_file.write_bytes(response.content)
+            hocr_file.write_bytes(self.hocr_data)
             self.assertLess(1, hocr_file.stat().st_size)
 
-        response = session.get(f"{self.BASE_URL}/max/{self.WORK}.jpg")
-        self.assertEqual(200, response.status_code)
-        jpg_file.write_bytes(response.content)
+        jpg_file.write_bytes(self.jpg_data)
         self.assertLess(1, jpg_file.stat().st_size)
 
     def _check_content(self, pdf_path: Path, key: str) -> None:
@@ -45,7 +54,7 @@ class HocrPdfTestCase(TestCase):
     def test_export_pdf(self) -> None:
         with TemporaryDirectory() as temp_directory:
             directory = Path(temp_directory)
-            self._download_files(directory)
+            self._generate_files(directory)
 
             pdf_path = directory / f"{self.WORK}.pdf"
             hocr_pdf.export_pdf(directory=str(directory), savefile=str(pdf_path))
@@ -54,7 +63,7 @@ class HocrPdfTestCase(TestCase):
     def test_export_pdf__kraken(self) -> None:
         with TemporaryDirectory() as temp_directory:
             directory = Path(temp_directory)
-            self._download_files(directory, skip_hocr=True)
+            self._generate_files(directory, skip_hocr=True)
 
             hocr_source_path = self.get_data_file("hocr_pdf/kraken.hocr")
             hocr_path = directory / f"{self.WORK}.hocr"
@@ -68,7 +77,7 @@ class HocrPdfTestCase(TestCase):
     def test_export_pdf__page_size(self) -> None:
         with TemporaryDirectory() as temp_directory:
             directory = Path(temp_directory)
-            self._download_files(directory)
+            self._generate_files(directory)
 
             image_path = directory / f"{self.WORK}.jpg"
             image: Image.Image = Image.new(mode="RGB", size=(2458, 3150), color=(0, 0, 0))
@@ -85,7 +94,7 @@ class HocrPdfTestCase(TestCase):
     def test_main(self) -> None:
         with TemporaryDirectory() as temp_directory:
             directory = Path(temp_directory)
-            self._download_files(directory)
+            self._generate_files(directory)
 
             with mock.patch("sys.argv", ["hocr-pdf", str(directory), "--savefile", str(directory / "output.pdf")]):
                 hocr_pdf.main()