+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Allow setting the maximum image pixel size for *Pillow* in `hocr_pdf`.
 * Drop support for Python <= 3.9.
+* Avoid using deprecated `argparse.FileType`.
 
 # Version 1.2.0 - 2025-07-01